requirements.txt
python-dotenv>=1.0.0
streamlit==1.45
redis